// src/app/core/model/address.ts
/**
 * O Alvo API - OpenAPI 3.0
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.3
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class Address { 
    zipcode?: string;
    street?: string;
    detail?: string;
    district?: string;
    city?: string;
    state?: string;
    country?: string;
}